+++ spec/rubocop/cop/capybara/rspec/match_style_spec.rb
@@ -1,17 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe RuboCop::Cop::Capybara::MatchStyle do
-  it 'registers an offense when using `assert_style`' do
-    expect_offense(<<~RUBY)
-      page.find(:css, '#first').assert_style(display: 'block')
-                                ^^^^^^^^^^^^ Use `assert_matches_style` instead of `assert_style`.
-    RUBY
-
-    expect_correction(<<~RUBY)
-      page.find(:css, '#first').assert_matches_style(display: 'block')
-    RUBY
-  end
-
+RSpec.describe RuboCop::Cop::Capybara::RSpec::MatchStyle do
   it 'registers an offense when using `has_style?`' do
     expect_offense(<<~RUBY)
       expect(page.find(:css, 'first')
@@ -33,12 +22,6 @@ RSpec.describe RuboCop::Cop::Capybara::MatchStyle do
 
     expect_correction(<<~RUBY)
       expect(page).to match_style(display: 'block')
-    RUBY
-  end
-
-  it 'does not register an offense when using `assert_matches_style`' do
-    expect_no_offenses(<<~RUBY)
-      page.find(:css, '#first').assert_matches_style(display: 'block')
     RUBY
   end